backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6

# ML dependencies
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# File processing
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# Report generation
reportlab==4.0.7

# Production/Config dependencies
pydantic-settings==2.1.0
slowapi==0.1.9
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0

requests>=2.31.0  # For weather API
